+++ tests/test_summary.py
@@ -68,6 +68,8 @@ def test_get_item(summary):
 
 def test_delete(summary):
     summary.update({"foo": "bar", "bad": True})
+    assert summary['foo'] == 'bar'
+    assert summary['bad'] is True
     del summary["bad"]
     assert disk_summary(summary) == {"foo": "bar"}
 
@@ -89,6 +91,7 @@ def test_matplot_image(summary):
 
 
 def test_matplot_plotly(summary):
+    plt.cla()
     plt.plot([1, 2, 3])
     summary["plot"] = plt
     plt.close()
@@ -120,6 +123,8 @@ def test_big_numpy(summary):
 def test_big_nested_numpy(summary):
     summary.update({"rad": {"deep": np.random.rand(1000)}})
     assert disk_summary(summary)["rad"]["deep"]["max"] > 0
+    summary["rad"]["deep2"] = np.random.rand(1000)
+    assert disk_summary(summary)["rad"]["deep2"]["max"] > 0
     assert os.path.exists(os.path.join(summary._run.dir, "wandb.h5"))
 
 
@@ -137,7 +142,8 @@ def test_tensorflow_tensor(summary):
 
 
 def test_pandas(summary):
-    summary.update({"pandas": pandas.DataFrame(data=np.random.rand(1000), columns=['col'])})
+    summary.update({"pandas": pandas.DataFrame(
+        data=np.random.rand(1000), columns=['col'])})
 
 
 def test_read_numpy(summary):
@@ -150,3 +156,21 @@ def test_read_nested_numpy(summary):
     summary.update({"rad": {"deep": np.random.rand(1000)}})
     s = FileSummary(summary._run)
     assert len(s["rad"]["deep"]) == 1000
+
+
+def test_read_very_nested_numpy(summary):
+    # Test that even deeply nested writes are written to disk.
+    summary.update(
+        {"a": {"b": {"c": {"d": {"e": {"f": {"g": {"h": {"i": {}}}}}}}}}})
+    summary['a']['b']['c']['d']['e']['f']['g']['h']['i']['j'] = True
+    assert disk_summary(summary)[
+        'a']['b']['c']['d']['e']['f']['g']['h']['i']['j'] is True
+
+
+def test_key_locking(summary):
+    summary.update({'a': 'a'})
+    assert summary['a'] == 'a'
+    summary.update({'a': 'b'})
+    assert summary['a'] == 'b'
+    summary.update({'a': 'c'}, overwrite=False)
+    assert summary['a'] == 'b'